fix: separating types & entities
