+++ src/components/AccountTypeSelect.tsx
@@ -1,11 +1,7 @@
 import { useFormContext } from "react-hook-form"
+import { TInputProps } from "@/types/accountTypeSelect/TInputProps"
 
-type inputProps = {
-    label: string,
-    id: string,
-}
-
-export default function({label,id}: inputProps){
+export default function({label,id}: TInputProps){
     const {register} = useFormContext()
     return <div className="flex flex-col gap-2">
         <label htmlFor={id} className="font-semibold text-sm" >{label}</label>
