+++ src/components/FooterText.tsx
@@ -1,12 +1,7 @@
 import Link from "next/link"
+import { TFooterTextProps } from "@/types/footerText/TFooterTextProps"
 
-type FooterTextProps = {
-  title: string,
-  firstTopic: string,
-  secondTopic: string | null
-}
-
-export default function FooterText( {title, firstTopic, secondTopic}: FooterTextProps ) {
+export default function FooterText( {title, firstTopic, secondTopic}: TFooterTextProps ) {
   return (
     <div className='flex justify-center flex-col'>
       <h1 className='mb-3 text-base'>{title}</h1>
