+++ src/components/Header.tsx
@@ -1,5 +1,3 @@
-
-
 export default function Header() {
     return (
         <div className="w-[100%] h-[50px] px-32 py-4 bg-blue-400 justify-between items-start inline-flex">
