+++ src/components/MaskFormInput.tsx
@@ -1,15 +1,8 @@
 import { Controller, useFormContext } from "react-hook-form"
 import { IMaskInput } from "react-imask"
+import { TInputProps } from "@/types/maskFormInput/TInputProps"
 
-type inputProps = {
-    label:string,
-    placeholder:string | undefined,
-    id:string,
-    mask?:string
-    classNameInput?:string | undefined
-}
-
-export default function({classNameInput,label,placeholder,id, mask}:inputProps){
+export default function({classNameInput,label,placeholder,id, mask}:TInputProps){
     const { control } = useFormContext()
     return <div className="flex flex-col gap-2">
     <label htmlFor={id} className={`font-semibold text-sm ${classNameInput}`} >{label}</label>
