+++ src/components/PrimaryButton.tsx
@@ -1,15 +1,8 @@
 import Link from "next/link"
 import { MouseEventHandler, ReactNode } from "react"
+import { TButtonProps } from "@/types/primaryButton/TButtonProps"
 
-type buttonProps = {
-    text:string,
-    className?:string | undefined,
-    onClick?:MouseEventHandler,
-    type?:"button" | "submit" | "reset" | undefined
-    href?: string
-}
-
-export default function({type="button", text, className, onClick, href}:buttonProps){
+export default function({type="button", text, className, onClick, href}:TButtonProps){
     return<>
         { href ? (
             <Link href={href}>
