+++ src/components/SectionForm.tsx
@@ -1,18 +1,12 @@
 "use client"
-
-import { ReactNode, useState } from "react"
+import { useState } from "react"
 import PrimaryButton from "./PrimaryButton"
 import SecondaryButton from "./SecondaryButton"
 import SectionIndicator from "./SectionIndicator"
-import { FieldValues, FormProvider, useForm } from "react-hook-form"
+import { FormProvider, useForm } from "react-hook-form"
+import { TFormProps } from "@/types/sectionForm/TFormProps"
 
-type formProps = {
-    className:string
-    sections:Array<Array<ReactNode>>
-    onFetch:(data:FieldValues) => void
-}
-
-export default function({className,sections, onFetch}:formProps){
+export default function({className,sections, onFetch}:TFormProps){
     const [currentSection,setCurrentSection] = useState(0)
     const methods = useForm()
