+++ src/entities/Book.ts
@@ -1,0 +1,22 @@
+import { EStatus } from "@/enums/EStatus"
+
+export type Book = {
+    author: string 
+    cover_url: string,
+    date: string,
+    edition: number,
+    gender: string,
+    id: number,
+    id_producer: number,
+    language: string,
+    name: string,
+    pages_number: number,
+    pdf_url: string,
+    price: number,
+    publisher: string,
+    rating: number,
+    submission_date: string,
+    submission_reason: string,
+    submission_status: EStatus,
+    synopsis: string
+}
