+++ src/enums/EStatus.ts
@@ -1,0 +1,5 @@
+export enum EStatus {
+    approved = 'Aprovado',
+    disapproved = 'Reprovado',
+    pending = 'Pendente'
+}
