+++ src/types/accountTypeSelect/TInputProps.ts
@@ -1,0 +1,4 @@
+export type TInputProps = {
+    label: string,
+    id: string,
+}
