+++ src/types/blueField/TElementsField.ts
@@ -1,0 +1,5 @@
+export type TElementsField = {
+    title: string,
+    value: string,
+    variation:string
+}
