+++ src/types/bookCarousel/TBookCarouselProps.ts
@@ -1,0 +1,9 @@
+export type TBookCarouselProps = {
+    id: number | undefined,
+    img: string | undefined,
+    title: string | undefined,
+    author: string | undefined,
+    currentPrice: number | undefined,
+    functionality?: () => void,
+    href: any
+  }
