+++ src/types/bookCover/TBookCoverProps.ts
@@ -1,0 +1,5 @@
+export type TBookCoverProps = {
+    img: string | undefined,
+    title: string | undefined,
+    autor: string | undefined,
+}
