+++ src/types/booksCarousel/TBooksCarrousel.ts
@@ -1,0 +1,6 @@
+import { Book } from "@/app/lib/page"
+
+export type TBooksCarrousel = {
+    title: string
+    booksList: Book[]
+}
