+++ src/types/buyCard/TCardProps.ts
@@ -1,0 +1,10 @@
+export type TCardProps = {
+    bookName: string | undefined,
+    author: string | undefined,
+    currentPrice: number | undefined,
+    textBtn: string | undefined,
+    functionality: () => void,
+    btnText: string
+    changeBtnClas: string,
+    hrefDown: string | null
+}
