+++ src/types/downloadButton/TButtonProps.ts
@@ -1,0 +1,6 @@
+export type TButtonProps = {
+    className?:string | undefined
+    functionality?: () => void
+    btnText: string,
+    hrefDown: string | null
+}
