+++ src/types/footerText/TFooterTextProps.ts
@@ -1,0 +1,5 @@
+export type TFooterTextProps = {
+    title: string,
+    firstTopic: string,
+    secondTopic: string | null
+}
