+++ src/types/inputRadio/TRadioProps.ts
@@ -1,0 +1,11 @@
+import { MouseEventHandler } from "react"
+
+export type TRadioProps = {
+    onClick?:MouseEventHandler<HTMLInputElement>
+    name:string,
+    id:string,
+    value:string,
+    label:string,
+    className?:string
+    checked?:boolean
+}
