+++ src/types/maskFormInput/TInputProps.ts
@@ -1,0 +1,7 @@
+export type TInputProps = {
+    label:string,
+    placeholder:string | undefined,
+    id:string,
+    mask?:string
+    classNameInput?:string | undefined
+}
