+++ src/types/orderCard/TOrderCardProps.ts
@@ -1,0 +1,8 @@
+export type TOrderCardProps = {
+    id: number | undefined,
+    img: string | undefined,
+    title: string | undefined,
+    author: string | undefined,
+    currentPrice: number | undefined,
+    date: string | undefined
+}
