+++ src/types/secondaryButton/TButtonProps.ts
@@ -1,0 +1,7 @@
+import { MouseEventHandler } from "react"
+
+export type TButtonProps = {
+    text:string,
+    className?:string | undefined,
+    onClick?:MouseEventHandler
+}
