+++ src/types/sectionIndicator/TIndicatorProps.ts
@@ -1,0 +1,6 @@
+import { statusName } from "./TStatusName"
+
+export type TIndicatorProps = {
+    status: statusName
+    index:number
+}
