+++ src/types/sectionIndicator/TStatusName.ts
@@ -1,0 +1,1 @@
+export type statusName = "current" | "completed" | "next"
