+++ src/types/specification/TSpecificationProps.ts
@@ -1,0 +1,8 @@
+export type SpecificationProps = {
+    type: string,
+    author: string | undefined,
+    pagesNumber: number | undefined,
+    category: string | undefined,
+    publishCompany: string | undefined,
+    date: string | undefined
+}
